requirements.txt
aiohttp==3.6.2
argon2-cffi==19.2.0
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
discord.py @ https://github.com/Rapptz/discord.py/archive/master.zip
Django==2.2.10
django-allauth==0.41.0
django-bootstrap3==12.0.3
django-braces==1.14.0
django-cleanup==5.0.0
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-crispy-forms==1.8.1
django-cron==0.5.1
django-filter==2.3.0
django-model-utils==4.0.0
django-private-chat==0.2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
httplib2==0.18.1
idna==2.7
idna-ssl==1.1.0
iniconfig==1.1.1
Jinja2==2.10.3
linkedin-api==1.1.0
MarkupSafe==1.1.1
multidict==4.7.6
oauth2client==3.0.0
oauthlib==3.1.0
packaging==20.4
parsedatetime==2.5
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
python3-openid==3.2.0
pytz==2019.3
requests==2.20.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.24.3
websockets==8.1
yarl==1.4.2
aiohttp==3.6.2
argon2-cffi==19.2.0
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coverage==5.3
coveralls==2.1.2
decorator==4.4.1
defusedxml==0.6.0
discord.py @ https://github.com/Rapptz/discord.py/archive/master.zip
Django==2.2.10
django-allauth==0.41.0
django-bootstrap3==12.0.3
django-braces==1.14.0
django-cleanup==5.0.0
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-crispy-forms==1.8.1
django-cron==0.5.1
django-filter==2.3.0
django-model-utils==4.0.0
django-private-chat==0.2.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
docopt==0.6.2
httplib2==0.18.1
idna==2.7
idna-ssl==1.1.0
iniconfig==1.1.1
Jinja2==2.10.3
linkedin-api==1.1.0
MarkupSafe==1.1.1
multidict==4.7.6
oauth2client==3.0.0
oauthlib==3.1.0
packaging==20.4
parsedatetime==2.5
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
python3-openid==3.2.0
pytz==2019.3
requests==2.20.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.24.3
websockets==8.1
yarl==1.4.2